+++ backend/dataall/modules/dataset_sharing/services/share_object_service.py
@@ -19,7 +19,6 @@ from dataall.modules.dataset_sharing.services.dataset_sharing_enums import (
     ShareItemStatus,
     ShareObjectStatus,
     PrincipalType,
-    ShareItemHealthStatus,
 )
 from dataall.modules.dataset_sharing.db.share_object_models import ShareObjectItem, ShareObject
 from dataall.modules.dataset_sharing.db.share_object_repositories import (
@@ -41,10 +40,7 @@ from dataall.modules.dataset_sharing.services.share_permissions import (
     DELETE_SHARE_OBJECT,
     GET_SHARE_OBJECT,
 )
-from dataall.modules.datasets.services.dataset_permissions import (
-    MANAGE_DATASETS,
-    UPDATE_DATASET,
-)
+
 from dataall.modules.dataset_sharing.aws.glue_client import GlueClient
 from dataall.modules.datasets.db.dataset_repositories import DatasetRepository
 from dataall.modules.datasets.db.dataset_models import DatasetTable, Dataset, DatasetStorageLocation
@@ -569,19 +565,3 @@ class ShareObjectService:
                 resource_uri=location.itemUri,
                 resource_type=DatasetStorageLocation.__name__,
             )
-
-    @staticmethod
-    @TenantPolicyService.has_tenant_permission(MANAGE_DATASETS)
-    @ResourcePolicyService.has_resource_permission(UPDATE_DATASET)
-    def verify_dataset_share_objects(uri: str, share_uris: list):
-        #TODO: when we abstract dataset_sharing_base from s3_dataset_sharing this function won't pollute the ShareObject service
-        with get_context().db_engine.scoped_session() as session:
-            for share_uri in share_uris:
-                share = ShareObjectRepository.get_share_by_uri(session, share_uri)
-                states = ShareItemSM.get_share_item_revokable_states()
-                items = ShareObjectRepository.list_shareable_items(
-                    session, share, states, {'pageSize': 1000, 'isShared': True}
-                )
-                item_uris = [item.shareItemUri for item in items.get('nodes', [])]
-                ShareItemService.verify_items_share_object(uri=share_uri, item_uris=item_uris)
-        return True
