+++ backend/dataall/modules/datasets/aws/kms_dataset_client.py
@@ -2,7 +2,6 @@ import logging
 
 from dataall.base.aws.sts import SessionHelper
 from botocore.exceptions import ClientError
-from dataall.modules.dataset_sharing.aws.share_policy_verifier import SharePolicyVerifier
 
 
 log = logging.getLogger(__name__)
